[MODIFY] account와 관련된 api 코드를 한 폴더에 모음
